Added CategoryReducer
